+++ CHANGELOG.md
@@ -8,10 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Web-based overlay with current song ([#22]).
+- Player will no longer pause the current song (if it's playing) and will instead synchronize the state of the player with Spotify ([#18]).
 
 ### Changed
 - Cleaned up old cruft in the codebase (`gfx` module).
+- Moved log configuration to external file (see [example log4rs.yaml]).
 
+[example log4rs.yaml]: log4rs.yaml
+
+[#18]: https://github.com/udoprog/setmod/issues/18
 [#22]: https://github.com/udoprog/setmod/issues/22
 
 [Unreleased]: https://github.com/udoprog/setmod/compare/0.2.3...HEAD